refact: move endpoints to new file
