+++ src-tauri/src/server/endpoints/mod.rs
@@ -1,0 +1,5 @@
+pub mod playback;
+pub mod playlist;
+pub mod soundboard;
+pub mod sounds;
+pub mod tracks;
